phi-2/requirements.txt
torch
torchvision
torchaudio
datasets
evaluate
gdown
python-docx
peft
pandas
scikit-learn
sentence-transformers
transformers
bm25s